Added more buttons and their associated plugins (WIP)
Forecolor needs a lot of UI work to create the button so it is disabled
for now. Charmap seems to have some bad interaction with the focus code
so it always inserts symbols at the start of the editor. There is also
an artifact of the TinyMCE popups showing up as a dot to the top left of
the editor.