Try lighter toggled sidebar, tab indicator
This is an attempt to increase the connection between the toggled cog and the sidebar, as well as visually lightening the toggled cog — the dark gray can ask for more attention, perhaps, than it deserves.

This is an experiment, and a work in progress.

The X button is also hidden on the desktop, as the cog does the same. This needs to be considered also, if we like this approach. Finally, if this is to ever become a thing, the hover styles need love too, and there are a couple of @todo's that need addressing.